improve project layout
